georiva/requirements.txt
wagtail==7.2.1
dj-database-url==3.0.1
django-timescaledb==0.2.13
django-environ==0.12.0
psycopg2==2.9.11
django-dbbackup==5.0.1
django-countries==8.2.0
django-countries-geoextent>=0.0.1
django-cleanup==9.0.0
django-timezone-field>=7.1
django-redis==6.0.0
celery-singleton==0.3.1
django-celery-beat==2.8.1
django-celery-results==2.6.0
django-extensions==4.1
channels
daphne
minio==7.2.20
gunicorn==23.0.0
uvicorn[standard]==0.38.0
wagtail-font-awesome-svg==2.0
wagtail-icon-chooser>=0.3.0
python-dateutil==2.9.0.post0
wagtailgeowidget==9.1.0
django-polymorphic==4.1.0
django-vue-utils==0.1.8
django-deep-translator==1.5.2
drf-spectacular
pydantic==2.12.5
xarray==2025.11.0
django-storages==1.14.6
google-auth==2.43.0
google-cloud-storage==3.6.0
boto3==1.41.5
dask[complete]